weather_classification/requirements.txt
absl-py==1.4.0
asn1crypto==0.24.0
beautifulsoup4==4.6.0
blinker==1.4
cachetools==5.3.0
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==3.0.1
cryptography==2.1.4
cupshelpers==1.0
decorator==4.1.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
feedparser==5.2.1
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
guacamole==0.9.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==6.0.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
louis==3.5.0
macaroonbakery==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.2
nose==1.3.7
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
olefile==0.45.1
opencv-python==4.7.0.68
padme==1.1.1
pexpect==4.2.1
Pillow==9.4.0
pipp==0.0.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
six==1.11.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.4.0
ufw==0.36
urllib3==1.22
webencodings==0.5
Werkzeug==2.2.2
xkit==0.0.0
XlsxWriter==0.9.6
zipp==3.12.0
